+++ scripts/test-all.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+bash scripts/test.sh
+bash scripts/test-examples.sh