+++ scripts/test-examples.sh
@@ -18,5 +18,6 @@ if [[ "$OSTYPE" == "linux-gnu" ]]; then
     run_example "token"
 
     run_example "derive/metadata"
+    run_example "derive/record_metadata"
     run_example "derive/print"
 fi